temp_humidity_app/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1
django-ninja==1.2.2
django-sse==0.4.1
mysqlclient==2.2.4
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.1
typing_extensions==4.12.2
